Finished front-end functionality
Now everything behaves as it should behave, except some bugs
like it is not possible to add the 14th element of search results
to 'issues in work' list. Don't know why.
I've also refactored the code a little.
And still we need to add a feature to store manager results
in local storage.