A complete new version with server
Now project has server and client side parts.
This was done mainly for educational purposes.
But still this version allows us some back end code. Which we use to
store manager data in json file and not in local.storage as before.
And we use it for many other purposes too. Like rendering html code
from pug templates and many others.